services/retrieval_service/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
numpy
pydantic